requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Cython==3.0.12
Django==5.1.7
django-cors-headers==4.7.0
django-filter==23.5
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
Markdown==3.7
MarkupSafe==3.0.2
mercadopago==2.3.0
pillow==11.1.0
psutil==7.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyngrok==7.2.3
pynput==1.8.1
pyperclip==1.9.0
python-decouple==3.8
python-dotenv==1.1.0
pywin32==310
PyYAML==6.0.2
ratelimit==2.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
